Cleanup: 0 warnings!

+++ FoolsEngine/src/FoolsEngine/Platform/Win10/Win10InputPolling.cpp
@@ -29,7 +29,7 @@ namespace fe {
         GLFWwindow* window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
         double x, y;
         glfwGetCursorPos(window, &x, &y);
-        return { x, y };
+        return { (float)x, (float)y };
     }
 
     float Win10InputPolling::GetMouseXNative()
